+++ pkg/kubelet/cm/BUILD
@@ -50,7 +50,7 @@ go_library(
             "//pkg/apis/core/v1/helper:go_default_library",
             "//pkg/apis/core/v1/helper/qos:go_default_library",
             "//pkg/kubelet/cadvisor:go_default_library",
-            "//pkg/kubelet/cm/cpumanager/containermap:go_default_library",
+            "//pkg/kubelet/cm/containermap:go_default_library",
             "//pkg/kubelet/cm/cpumanager/topology:go_default_library",
             "//pkg/kubelet/cm/devicemanager:go_default_library",
             "//pkg/kubelet/cm/util:go_default_library",
@@ -101,7 +101,7 @@ go_library(
             "//pkg/apis/core/v1/helper:go_default_library",
             "//pkg/apis/core/v1/helper/qos:go_default_library",
             "//pkg/kubelet/cadvisor:go_default_library",
-            "//pkg/kubelet/cm/cpumanager/containermap:go_default_library",
+            "//pkg/kubelet/cm/containermap:go_default_library",
             "//pkg/kubelet/cm/cpumanager/topology:go_default_library",
             "//pkg/kubelet/cm/devicemanager:go_default_library",
             "//pkg/kubelet/cm/util:go_default_library",
@@ -212,6 +212,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//pkg/kubelet/cm/containermap:all-srcs",
         "//pkg/kubelet/cm/cpumanager:all-srcs",
         "//pkg/kubelet/cm/cpuset:all-srcs",
         "//pkg/kubelet/cm/devicemanager:all-srcs",
