+++ pkg/kubelet/cm/cpumanager/BUILD
@@ -14,7 +14,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/core/v1/helper/qos:go_default_library",
-        "//pkg/kubelet/cm/cpumanager/containermap:go_default_library",
+        "//pkg/kubelet/cm/containermap:go_default_library",
         "//pkg/kubelet/cm/cpumanager/state:go_default_library",
         "//pkg/kubelet/cm/cpumanager/topology:go_default_library",
         "//pkg/kubelet/cm/cpuset:go_default_library",
@@ -42,7 +42,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/kubelet/cm/cpumanager/containermap:go_default_library",
+        "//pkg/kubelet/cm/containermap:go_default_library",
         "//pkg/kubelet/cm/cpumanager/state:go_default_library",
         "//pkg/kubelet/cm/cpumanager/topology:go_default_library",
         "//pkg/kubelet/cm/cpuset:go_default_library",
@@ -68,7 +68,6 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//pkg/kubelet/cm/cpumanager/containermap:all-srcs",
         "//pkg/kubelet/cm/cpumanager/state:all-srcs",
         "//pkg/kubelet/cm/cpumanager/topology:all-srcs",
     ],
