+++ pkg/kubelet/cm/cpumanager/state/state_file.go
@@ -24,7 +24,7 @@ import (
 	"sync"
 
 	"k8s.io/klog"
-	"k8s.io/kubernetes/pkg/kubelet/cm/cpumanager/containermap"
+	"k8s.io/kubernetes/pkg/kubelet/cm/containermap"
 	"k8s.io/kubernetes/pkg/kubelet/cm/cpuset"
 )
